requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
anthropic>=0.18.0
pydantic>=2.6.0
python-dotenv>=1.0.0
websockets>=12.0
aiofiles>=23.2.0
faster-whisper>=1.0.0
